Minor change
From the January 2016 issue, MDPI journals will use article numbers in place of the traditional method of continuous pagination through the volume. This step helps us to maintain a rapid, efficient production process by being able to define pagination as soon as a paper is accepted. We have recently made a number of improvements to article layout and production procedures, of which article numbers is just one component.